[FIX] fix @lmignon comment
