+++ shopinvader_payment/services/__init__.py
@@ -1,2 +1,3 @@
+from . import abstract_payable_sale
 from . import cart
 from . import invader_payment_service
