+++ shopinvader_quotation_payment/services/cart.py
@@ -13,5 +13,9 @@ class CartService(Component):
     def _get_available_payment_mode(self, cart):
         for line in cart.order_line:
             if line.product_id.only_quotation:
+                # If we have a product that required a quotation in the cart
+                # (product without public price) we remove all payment
+                # method as it's not possible to pay it.
+                # Customer must request a quotation for this cart
                 return self.env["shopinvader.payment"].browse(False)
         return super(CartService, self)._get_available_payment_mode(cart)
